requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
youtube-transcript-api
feedparser
requests
beautifulsoup4
pytube
google-generativeai
yt-dlp
# 離線語音轉文字備援所需套件（可選，但若要啟用 fallback 請務必安裝）
# pip install --upgrade openai-whisper
# 需要系統層級安裝 ffmpeg，可搭配 Python 介面：
# pip install ffmpeg-python
